Add terminal pages
