+++ src/pages/HomePage.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import FloatingCharacters from '../components/FloatingCharacters';
+import MenuOverlay from '../components/MenuOverlay';
+
+const HomePage = () => {
+    return (
+        <>
+            <FloatingCharacters />
+            <MenuOverlay />
+        </>
+    );
+};
+
+export default HomePage;
